chore: Update CoreCrypto to v10.5.2 #WPB-28387
* bump core-crypto-jvm to 10.5.2
* migrate MLS crypto client code to the new CoreCrypto API types and transaction flow
* replace string-based CryptoClientId handling with structured user/device/domain fields
* update decrypted MLS messages to carry a QualifiedId sender and allow non-text decrypt results
* serialize key packages and welcomes with new CoreCrypto serializers instead of raw copied bytes
* adapt MLS transport to the new MlsTransport contract and keep commit upload failures non-fatal
* remove sender-client-id string parsing from EventsRouter
* add explicit error handling for MLS operations used before credential initialization
* update service/router logic and tests for new MLS join/welcome return types
* refresh test fixtures to use valid hex device ids, valid MLS group info, and generated key packages
* add coverage for orphan welcome recovery and external commit bundle generation
* update affected unit/integration tests for the new CoreCrypto behavior

+++ lib/src/main/kotlin/com/wire/sdk/model/http/client/PreKeyCrypto.kt
@@ -17,8 +17,8 @@
 package com.wire.sdk.model.http.client
 
 data class PreKeyCrypto(
-    val id: Int,
+    val id: UShort,
     val encodedData: String
 )
 
-fun PreKeyCrypto.toApi(): PreKeyRequest = PreKeyRequest(id, encodedData)
+fun PreKeyCrypto.toApi(): PreKeyRequest = PreKeyRequest(id.toInt(), encodedData)
